added some demo code to the frontend
