+++ backend/api/article/models/article.settings.json
@@ -2,7 +2,8 @@
   "kind": "collectionType",
   "collectionName": "articles",
   "info": {
-    "name": "article"
+    "name": "article",
+    "description": ""
   },
   "options": {
     "increments": true,
@@ -22,6 +23,10 @@
     "topic": {
       "model": "topic",
       "via": "articles"
+    },
+    "about": {
+      "type": "text",
+      "required": true
     }
   }
 }
