+++ frontend/src/pages/Article.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const Article = () => {
+  return <div>article</div>;
+};
+
+export default Article;
